requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
certifi=2016.2.28=py36_0
click=6.7=py36_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.10.0=py36_0
dataclasses=0.8=pyh4f3eec9_6
flask=0.12.2=py36_0
freetype=2.10.4=hd328e21_0
icu=58.2=ha925a31_3
intel-openmp=2022.0.0=haa95532_3663
itsdangerous=0.24=py36_0
jinja2=2.9.6=py36_0
jpeg=9d=h2bbff1b_0
kiwisolver=1.3.1=py36hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=1.0=py36_0
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py36hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.44=py36_0
opencv-python=4.5.5.64=pypi_0
openssl=1.0.1k=1
pillow=8.3.1=py36h4fa10fc_0
pip=21.3.1=pypi_0
pyparsing=2.2.0=py36_0
pyqt=5.6.0=py36_2
python=3.6.13=h3758d61_0
python-dateutil=2.6.1=py36_0
pytorch=1.10.2=py3.6_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
qt=5.6.2=vc14h6f8c307_12
setuptools=36.4.0=py36_1
sip=4.18=py36_0
six=1.10.0=py36_0
sqlite=3.38.2=h2bbff1b_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.10.2=py36_cu113
torchvision=0.11.3=py36_cu113
tornado=4.5.2=py36_0
typing_extensions=4.1.1=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.12.2=py36_0
wheel=0.29.0=py36_0
wincertstore=0.2=py36_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
